requirements_final.txt
# 전체 시스템 통합 패키지 (모든 기능 포함)
# 모든 가상환경의 패키지를 통합한 완전한 시스템

# 웹 인터페이스
streamlit==1.47.0

# 데이터 처리
pandas==2.3.1
numpy==2.3.1

# PDF 처리
PyMuPDF==1.26.3
camelot-py==1.0.0
ghostscript==0.8.1
opencv-python-headless==4.12.0.88
pypdfium2==4.30.0
pdfminer.six==20250506
pypdf==3.17.4

# 시각화
plotly==6.2.0
altair==5.5.0
matplotlib

# LLM 처리
ollama==0.5.1
requests==2.32.4

# 임베딩 및 검색
sentence-transformers
faiss-cpu

# 자연어 처리
nltk==3.9.1
regex==2024.11.6

# 머신러닝
scikit-learn
pickle-mixin

# 유틸리티
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
idna==3.10
joblib==1.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0

# Streamlit 의존성
packaging==25.0
protobuf==6.31.1
pyarrow==21.0.0
pydantic==2.11.7
pydeck==0.9.1
watchdog==6.0.0

# 테이블 처리
openpyxl==3.1.5
tabulate==0.9.0
cffi==1.17.1
chardet==5.2.0
cryptography==45.0.5
et_xmlfile==2.0.0
pycparser==2.22
tzdata==2025.2

# 주요 기능:
# - PDF 텍스트/테이블 추출
# - LLM 섹션 제목/키워드/요약 추출
# - LLM 컬럼 추론 및 개선
# - 임베딩 벡터 생성
# - 웹 인터페이스
# - 결과 파일 다운로드

